LPS-171491 Extract trash services calls from portal-impl DLAppHelperLocalServiceImpl to wrapper class
