+++ portal-impl/src/com/liferay/portlet/documentlibrary/service/impl/DLAppHelperLocalServiceImpl.java
@@ -29,21 +29,12 @@ import com.liferay.document.library.kernel.model.DLFileShortcutConstants;
 import com.liferay.document.library.kernel.model.DLFileVersion;
 import com.liferay.document.library.kernel.model.DLFolder;
 import com.liferay.document.library.kernel.model.DLFolderConstants;
-import com.liferay.document.library.kernel.service.DLAppLocalService;
 import com.liferay.document.library.kernel.service.DLAppService;
 import com.liferay.document.library.kernel.service.DLFileEntryLocalService;
 import com.liferay.document.library.kernel.service.DLFileShortcutLocalService;
-import com.liferay.document.library.kernel.service.DLFileVersionLocalService;
-import com.liferay.document.library.kernel.service.DLFolderLocalService;
-import com.liferay.document.library.kernel.service.persistence.DLFileEntryFinder;
-import com.liferay.document.library.kernel.service.persistence.DLFileEntryPersistence;
 import com.liferay.document.library.kernel.service.persistence.DLFileShortcutPersistence;
 import com.liferay.document.library.kernel.service.persistence.DLFileVersionPersistence;
-import com.liferay.document.library.kernel.service.persistence.DLFolderFinder;
-import com.liferay.document.library.kernel.service.persistence.DLFolderPersistence;
 import com.liferay.document.library.kernel.util.DLAppHelperThreadLocal;
-import com.liferay.document.library.kernel.util.DLUtil;
-import com.liferay.document.library.kernel.util.comparator.DLFileVersionVersionComparator;
 import com.liferay.portal.kernel.bean.BeanReference;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.DynamicQuery;
@@ -53,18 +44,15 @@ import com.liferay.portal.kernel.dao.orm.ProjectionFactoryUtil;
 import com.liferay.portal.kernel.dao.orm.Property;
 import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
 import com.liferay.portal.kernel.dao.orm.RestrictionsFactoryUtil;
-import com.liferay.portal.kernel.dao.orm.WildcardMode;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.interval.IntervalActionProcessor;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
-import com.liferay.portal.kernel.lock.Lock;
 import com.liferay.portal.kernel.messaging.async.Async;
 import com.liferay.portal.kernel.repository.Repository;
 import com.liferay.portal.kernel.repository.RepositoryProviderUtil;
 import com.liferay.portal.kernel.repository.capabilities.RepositoryEventTriggerCapability;
 import com.liferay.portal.kernel.repository.event.RepositoryEventType;
-import com.liferay.portal.kernel.repository.event.TrashRepositoryEventType;
 import com.liferay.portal.kernel.repository.event.WorkflowRepositoryEventType;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.repository.model.FileShortcut;
@@ -73,44 +61,24 @@ import com.liferay.portal.kernel.repository.model.Folder;
 import com.liferay.portal.kernel.repository.model.RepositoryModel;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
-import com.liferay.portal.kernel.service.ClassNameLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.WorkflowInstanceLinkLocalService;
 import com.liferay.portal.kernel.social.SocialActivityManagerUtil;
-import com.liferay.portal.kernel.trash.helper.TrashHelper;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.ObjectValuePair;
 import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
-import com.liferay.portal.kernel.util.ServiceProxyFactory;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.UnicodePropertiesBuilder;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.repository.liferayrepository.model.LiferayFileEntry;
-import com.liferay.portal.repository.liferayrepository.model.LiferayFileShortcut;
-import com.liferay.portal.repository.liferayrepository.model.LiferayFileVersion;
 import com.liferay.portal.repository.liferayrepository.model.LiferayFolder;
 import com.liferay.portlet.documentlibrary.service.base.DLAppHelperLocalServiceBaseImpl;
 import com.liferay.portlet.documentlibrary.social.DLActivityKeys;
-import com.liferay.portlet.documentlibrary.util.DLAppUtil;
 import com.liferay.ratings.kernel.service.RatingsStatsLocalService;
-import com.liferay.social.kernel.model.SocialActivityConstants;
-import com.liferay.trash.kernel.exception.RestoreEntryException;
-import com.liferay.trash.kernel.exception.TrashEntryException;
-import com.liferay.trash.kernel.model.TrashEntry;
-import com.liferay.trash.kernel.model.TrashVersion;
-import com.liferay.trash.kernel.service.TrashEntryLocalService;
-import com.liferay.trash.kernel.service.TrashVersionLocalService;
-import com.liferay.util.dao.orm.CustomSQLUtil;
 
 import java.io.Serializable;
 
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -326,8 +294,6 @@ public class DLAppHelperLocalServiceImpl
 	@Override
 	public void moveDependentsToTrash(DLFolder dlFolder)
 		throws PortalException {
-
-		trashOrRestoreFolder(dlFolder, true);
 	}
 
 	@Override
@@ -336,24 +302,7 @@ public class DLAppHelperLocalServiceImpl
 			ServiceContext serviceContext)
 		throws PortalException {
 
-		boolean hasLock = _dlFileEntryLocalService.hasFileEntryLock(
-			userId, fileEntry.getFileEntryId(), fileEntry.getFolderId());
-
-		if (!hasLock) {
-			_dlFileEntryLocalService.lockFileEntry(
-				userId, fileEntry.getFileEntryId());
-		}
-
-		try {
-			return doMoveFileEntryFromTrash(
-				userId, fileEntry, newFolderId, serviceContext);
-		}
-		finally {
-			if (!hasLock) {
-				_dlFileEntryLocalService.unlockFileEntry(
-					fileEntry.getFileEntryId());
-			}
-		}
+		return null;
 	}
 
 	/**
@@ -367,28 +316,7 @@ public class DLAppHelperLocalServiceImpl
 	public FileEntry moveFileEntryToTrash(long userId, FileEntry fileEntry)
 		throws PortalException {
 
-		boolean hasLock = _dlFileEntryLocalService.hasFileEntryLock(
-			userId, fileEntry.getFileEntryId(), fileEntry.getFolderId());
-
-		if (!hasLock) {
-			_dlFileEntryLocalService.lockFileEntry(
-				userId, fileEntry.getFileEntryId());
-		}
-
-		try {
-			if (fileEntry.isCheckedOut()) {
-				_dlFileEntryLocalService.cancelCheckOut(
-					userId, fileEntry.getFileEntryId());
-			}
-
-			return doMoveFileEntryToTrash(userId, fileEntry);
-		}
-		finally {
-			if (!hasLock) {
-				_dlFileEntryLocalService.unlockFileEntry(
-					fileEntry.getFileEntryId());
-			}
-		}
+		return null;
 	}
 
 	@Override
@@ -397,54 +325,7 @@ public class DLAppHelperLocalServiceImpl
 			ServiceContext serviceContext)
 		throws PortalException {
 
-		DLFileShortcut dlFileShortcut = (DLFileShortcut)fileShortcut.getModel();
-
-		if (!dlFileShortcut.isInTrash()) {
-			throw new RestoreEntryException(
-				RestoreEntryException.INVALID_STATUS);
-		}
-
-		if (_trashHelper.isInTrashExplicitly(dlFileShortcut)) {
-			restoreFileShortcutFromTrash(userId, fileShortcut);
-		}
-		else {
-
-			// File shortcut
-
-			TrashVersion trashVersion = _trashVersionLocalService.fetchVersion(
-				DLFileShortcutConstants.getClassName(),
-				fileShortcut.getFileShortcutId());
-
-			int status = WorkflowConstants.STATUS_APPROVED;
-
-			if (trashVersion != null) {
-				status = trashVersion.getStatus();
-			}
-
-			_dlFileShortcutLocalService.updateStatus(
-				userId, fileShortcut.getFileShortcutId(), status,
-				new ServiceContext());
-
-			// Trash
-
-			if (trashVersion != null) {
-				_trashVersionLocalService.deleteTrashVersion(trashVersion);
-			}
-
-			// Social
-
-			JSONObject extraDataJSONObject = JSONUtil.put(
-				"title", fileShortcut.getToTitle());
-
-			SocialActivityManagerUtil.addActivity(
-				userId, fileShortcut,
-				SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
-				extraDataJSONObject.toString(), 0);
-		}
-
-		return _dlAppLocalService.updateFileShortcut(
-			userId, fileShortcut.getFileShortcutId(), newFolderId,
-			fileShortcut.getToFileEntryId(), serviceContext);
+		return null;
 	}
 
 	/**
@@ -459,38 +340,7 @@ public class DLAppHelperLocalServiceImpl
 			long userId, FileShortcut fileShortcut)
 		throws PortalException {
 
-		// File shortcut
-
-		DLFileShortcut dlFileShortcut = (DLFileShortcut)fileShortcut.getModel();
-
-		if (dlFileShortcut.isInTrash()) {
-			throw new TrashEntryException();
-		}
-
-		int oldStatus = dlFileShortcut.getStatus();
-
-		dlFileShortcut = _dlFileShortcutLocalService.updateStatus(
-			userId, fileShortcut.getFileShortcutId(),
-			WorkflowConstants.STATUS_IN_TRASH, new ServiceContext());
-
-		// Social
-
-		JSONObject extraDataJSONObject = JSONUtil.put(
-			"title", _trashHelper.getOriginalTitle(fileShortcut.getToTitle()));
-
-		SocialActivityManagerUtil.addActivity(
-			userId, fileShortcut, SocialActivityConstants.TYPE_MOVE_TO_TRASH,
-			extraDataJSONObject.toString(), 0);
-
-		// Trash
-
-		_trashEntryLocalService.addTrashEntry(
-			userId, fileShortcut.getGroupId(),
-			DLFileShortcutConstants.getClassName(),
-			fileShortcut.getFileShortcutId(), fileShortcut.getUuid(), null,
-			oldStatus, null, null);
-
-		return new LiferayFileShortcut(dlFileShortcut);
+		return null;
 	}
 
 	@Override
@@ -499,26 +349,7 @@ public class DLAppHelperLocalServiceImpl
 			ServiceContext serviceContext)
 		throws PortalException {
 
-		boolean hasLock = _dlFolderLocalService.hasFolderLock(
-			userId, folder.getFolderId());
-
-		Lock lock = null;
-
-		if (!hasLock) {
-			lock = _dlFolderLocalService.lockFolder(
-				userId, folder.getFolderId());
-		}
-
-		try {
-			return doMoveFolderFromTrash(
-				userId, folder, parentFolderId, serviceContext);
-		}
-		finally {
-			if (!hasLock) {
-				_dlFolderLocalService.unlockFolder(
-					folder.getFolderId(), lock.getUuid());
-			}
-		}
+		return null;
 	}
 
 	/**
@@ -532,25 +363,7 @@ public class DLAppHelperLocalServiceImpl
 	public Folder moveFolderToTrash(long userId, Folder folder)
 		throws PortalException {
 
-		boolean hasLock = _dlFolderLocalService.hasFolderLock(
-			userId, folder.getFolderId());
-
-		Lock lock = null;
-
-		if (!hasLock) {
-			lock = _dlFolderLocalService.lockFolder(
-				userId, folder.getFolderId());
-		}
-
-		try {
-			return doMoveFolderToTrash(userId, folder);
-		}
-		finally {
-			if (!hasLock) {
-				_dlFolderLocalService.unlockFolder(
-					folder.getFolderId(), lock.getUuid());
-			}
-		}
+		return null;
 	}
 
 	@Async
@@ -599,205 +412,28 @@ public class DLAppHelperLocalServiceImpl
 	@Override
 	public void restoreDependentsFromTrash(DLFolder dlFolder)
 		throws PortalException {
-
-		trashOrRestoreFolder(dlFolder, false);
 	}
 
 	@Override
 	public void restoreFileEntryFromTrash(long userId, FileEntry fileEntry)
 		throws PortalException {
-
-		restoreFileEntryFromTrash(userId, fileEntry.getFolderId(), fileEntry);
 	}
 
 	@Override
 	public void restoreFileEntryFromTrash(
 			long userId, long newFolderId, FileEntry fileEntry)
 		throws PortalException {
-
-		// File entry
-
-		DLFileEntry dlFileEntry = (DLFileEntry)fileEntry.getModel();
-
-		if (!dlFileEntry.isInTrash()) {
-			throw new RestoreEntryException(
-				RestoreEntryException.INVALID_STATUS);
-		}
-
-		FileVersion fileVersion = fileEntry.getFileVersion();
-
-		if (!DLAppHelperThreadLocal.isEnabled()) {
-			_dlFileEntryLocalService.updateStatus(
-				userId, fileVersion.getFileVersionId(),
-				WorkflowConstants.STATUS_APPROVED, new ServiceContext(),
-				new HashMap<String, Serializable>());
-
-			return;
-		}
-
-		String originalTitle = _trashHelper.getOriginalTitle(
-			dlFileEntry.getTitle());
-
-		String title = _dlFileEntryLocalService.getUniqueTitle(
-			dlFileEntry.getGroupId(), newFolderId, dlFileEntry.getFileEntryId(),
-			originalTitle, dlFileEntry.getExtension());
-
-		String originalFileName = _trashHelper.getOriginalTitle(
-			dlFileEntry.getTitle(), "fileName");
-
-		String fileName = originalFileName;
-
-		if (!StringUtil.equals(title, originalTitle)) {
-			fileName = DLUtil.getSanitizedFileName(
-				title, DLAppUtil.getExtension(title, originalFileName));
-		}
-
-		dlFileEntry.setFileName(fileName);
-		dlFileEntry.setTitle(title);
-
-		dlFileEntry = _dlFileEntryPersistence.update(dlFileEntry);
-
-		DLFileVersion dlFileVersion = (DLFileVersion)fileVersion.getModel();
-
-		dlFileVersion.setFileName(fileName);
-		dlFileVersion.setTitle(title);
-
-		dlFileVersion = _dlFileVersionPersistence.update(dlFileVersion);
-
-		TrashEntry trashEntry = _trashEntryLocalService.getEntry(
-			DLFileEntryConstants.getClassName(), fileEntry.getFileEntryId());
-
-		_dlFileEntryLocalService.updateStatus(
-			userId, dlFileEntry, dlFileVersion, trashEntry.getStatus(),
-			new ServiceContext(), new HashMap<>());
-
-		// File shortcut
-
-		_dlFileShortcutLocalService.enableFileShortcuts(
-			fileEntry.getFileEntryId());
-
-		// Sync
-
-		triggerRepositoryEvent(
-			fileEntry.getRepositoryId(),
-			TrashRepositoryEventType.EntryRestored.class, FileEntry.class,
-			fileEntry);
-
-		// Trash
-
-		List<TrashVersion> trashVersions =
-			_trashVersionLocalService.getVersions(trashEntry.getEntryId());
-
-		for (TrashVersion trashVersion : trashVersions) {
-			DLFileVersion trashDLFileVersion =
-				_dlFileVersionPersistence.findByPrimaryKey(
-					trashVersion.getClassPK());
-
-			trashDLFileVersion.setStatus(trashVersion.getStatus());
-
-			_dlFileVersionPersistence.update(trashDLFileVersion);
-		}
-
-		_trashEntryLocalService.deleteEntry(trashEntry.getEntryId());
-
-		// Social
-
-		JSONObject extraDataJSONObject = JSONUtil.put(
-			"title", fileEntry.getTitle());
-
-		SocialActivityManagerUtil.addActivity(
-			userId, fileEntry, SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
-			extraDataJSONObject.toString(), 0);
 	}
 
 	@Override
 	public void restoreFileShortcutFromTrash(
 			long userId, FileShortcut fileShortcut)
 		throws PortalException {
-
-		DLFileShortcut dlFileShortcut = (DLFileShortcut)fileShortcut.getModel();
-
-		if (!dlFileShortcut.isInTrash()) {
-			throw new RestoreEntryException(
-				RestoreEntryException.INVALID_STATUS);
-		}
-
-		// File shortcut
-
-		TrashEntry trashEntry = _trashEntryLocalService.getEntry(
-			DLFileShortcutConstants.getClassName(),
-			fileShortcut.getFileShortcutId());
-
-		_dlFileShortcutLocalService.updateStatus(
-			userId, fileShortcut.getFileShortcutId(), trashEntry.getStatus(),
-			new ServiceContext());
-
-		// Social
-
-		JSONObject extraDataJSONObject = JSONUtil.put(
-			"title", fileShortcut.getToTitle());
-
-		SocialActivityManagerUtil.addActivity(
-			userId, fileShortcut,
-			SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
-			extraDataJSONObject.toString(), 0);
-
-		// Trash
-
-		_trashEntryLocalService.deleteEntry(trashEntry.getEntryId());
 	}
 
 	@Override
 	public void restoreFolderFromTrash(long userId, Folder folder)
 		throws PortalException {
-
-		// Folder
-
-		DLFolder dlFolder = (DLFolder)folder.getModel();
-
-		if (!dlFolder.isInTrash()) {
-			throw new RestoreEntryException(
-				RestoreEntryException.INVALID_STATUS);
-		}
-
-		String originalName = _trashHelper.getOriginalTitle(dlFolder.getName());
-
-		dlFolder.setName(
-			_dlFolderLocalService.getUniqueFolderName(
-				folder.getUuid(), folder.getGroupId(),
-				folder.getParentFolderId(), originalName, 2));
-
-		dlFolder = _dlFolderPersistence.update(dlFolder);
-
-		TrashEntry trashEntry = _trashEntryLocalService.getEntry(
-			DLFolder.class.getName(), dlFolder.getFolderId());
-
-		_dlFolderLocalService.updateStatus(
-			userId, folder.getFolderId(), trashEntry.getStatus(),
-			new HashMap<String, Serializable>(), new ServiceContext());
-
-		// Folders, file entries, and file shortcuts
-
-		restoreDependentsFromTrash(dlFolder);
-
-		// Sync
-
-		triggerRepositoryEvent(
-			folder.getRepositoryId(),
-			TrashRepositoryEventType.EntryRestored.class, Folder.class, folder);
-
-		// Trash
-
-		_trashEntryLocalService.deleteEntry(trashEntry.getEntryId());
-
-		// Social
-
-		JSONObject extraDataJSONObject = JSONUtil.put(
-			"title", folder.getName());
-
-		SocialActivityManagerUtil.addActivity(
-			userId, folder, SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
-			extraDataJSONObject.toString(), 0);
 	}
 
 	@Override
@@ -1193,360 +829,6 @@ public class DLAppHelperLocalServiceImpl
 		}
 	}
 
-	protected FileEntry doMoveFileEntryFromTrash(
-			long userId, FileEntry fileEntry, long newFolderId,
-			ServiceContext serviceContext)
-		throws PortalException {
-
-		// File entry
-
-		DLFileEntry dlFileEntry = (DLFileEntry)fileEntry.getModel();
-
-		if (!dlFileEntry.isInTrash()) {
-			throw new RestoreEntryException(
-				RestoreEntryException.INVALID_STATUS);
-		}
-
-		if (_trashHelper.isInTrashExplicitly(dlFileEntry)) {
-			restoreFileEntryFromTrash(userId, newFolderId, fileEntry);
-
-			if (fileEntry.getFolderId() != newFolderId) {
-				fileEntry = _dlAppLocalService.moveFileEntry(
-					userId, fileEntry.getFileEntryId(), newFolderId,
-					serviceContext);
-			}
-
-			// Indexer
-
-			Indexer<DLFileEntry> indexer =
-				IndexerRegistryUtil.nullSafeGetIndexer(DLFileEntry.class);
-
-			indexer.reindex((DLFileEntry)fileEntry.getModel());
-
-			return fileEntry;
-		}
-
-		List<DLFileVersion> dlFileVersions =
-			_dlFileVersionLocalService.getFileVersions(
-				fileEntry.getFileEntryId(), WorkflowConstants.STATUS_IN_TRASH);
-
-		dlFileVersions = ListUtil.sort(
-			dlFileVersions, new DLFileVersionVersionComparator());
-
-		FileVersion fileVersion = new LiferayFileVersion(dlFileVersions.get(0));
-
-		TrashVersion trashVersion = _trashVersionLocalService.fetchVersion(
-			DLFileVersion.class.getName(), fileVersion.getFileVersionId());
-
-		int oldStatus = WorkflowConstants.STATUS_APPROVED;
-
-		if (trashVersion != null) {
-			oldStatus = trashVersion.getStatus();
-		}
-
-		_dlFileEntryLocalService.updateStatus(
-			userId, dlFileEntry, dlFileVersions.get(0), oldStatus,
-			serviceContext, new HashMap<>());
-
-		// File versions
-
-		for (DLFileVersion dlFileVersion : dlFileVersions) {
-
-			// File version
-
-			trashVersion = _trashVersionLocalService.fetchVersion(
-				DLFileVersion.class.getName(),
-				dlFileVersion.getFileVersionId());
-
-			oldStatus = WorkflowConstants.STATUS_APPROVED;
-
-			if (trashVersion != null) {
-				oldStatus = trashVersion.getStatus();
-			}
-
-			dlFileVersion.setStatus(oldStatus);
-
-			_dlFileVersionPersistence.update(dlFileVersion);
-
-			// Trash
-
-			if (trashVersion != null) {
-				_trashVersionLocalService.deleteTrashVersion(trashVersion);
-			}
-		}
-
-		if (DLAppHelperThreadLocal.isEnabled()) {
-
-			// File shortcut
-
-			_dlFileShortcutLocalService.enableFileShortcuts(
-				fileEntry.getFileEntryId());
-		}
-
-		// App helper
-
-		fileEntry = _dlAppLocalService.moveFileEntry(
-			userId, fileEntry.getFileEntryId(), newFolderId, serviceContext);
-
-		// Sync
-
-		triggerRepositoryEvent(
-			fileEntry.getRepositoryId(),
-			TrashRepositoryEventType.EntryRestored.class, FileEntry.class,
-			fileEntry);
-
-		// Social
-
-		JSONObject extraDataJSONObject = JSONUtil.put(
-			"title", fileEntry.getTitle());
-
-		SocialActivityManagerUtil.addActivity(
-			userId, fileEntry, SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
-			extraDataJSONObject.toString(), 0);
-
-		// Indexer
-
-		Indexer<DLFileEntry> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
-			DLFileEntry.class);
-
-		indexer.reindex((DLFileEntry)fileEntry.getModel());
-
-		return fileEntry;
-	}
-
-	protected FileEntry doMoveFileEntryToTrash(long userId, FileEntry fileEntry)
-		throws PortalException {
-
-		if (fileEntry.isInTrash()) {
-			throw new TrashEntryException();
-		}
-
-		List<DLFileVersion> dlFileVersions =
-			_dlFileVersionLocalService.getFileVersions(
-				fileEntry.getFileEntryId(), WorkflowConstants.STATUS_ANY);
-
-		dlFileVersions = ListUtil.sort(
-			dlFileVersions, new DLFileVersionVersionComparator());
-
-		List<ObjectValuePair<Long, Integer>> dlFileVersionStatusOVPs =
-			new ArrayList<>();
-
-		if ((dlFileVersions != null) && !dlFileVersions.isEmpty()) {
-			dlFileVersionStatusOVPs = getDlFileVersionStatuses(dlFileVersions);
-		}
-
-		FileVersion fileVersion = fileEntry.getLatestFileVersion(true);
-
-		_dlFileEntryLocalService.updateStatus(
-			userId, fileVersion.getFileVersionId(),
-			WorkflowConstants.STATUS_IN_TRASH, new ServiceContext(),
-			new HashMap<String, Serializable>());
-
-		if (DLAppHelperThreadLocal.isEnabled()) {
-
-			// File shortcut
-
-			_dlFileShortcutLocalService.disableFileShortcuts(
-				fileEntry.getFileEntryId());
-
-			// Sync
-
-			triggerRepositoryEvent(
-				fileEntry.getRepositoryId(),
-				TrashRepositoryEventType.EntryTrashed.class, FileEntry.class,
-				fileEntry);
-		}
-
-		// Trash
-
-		dlFileVersions = _dlFileVersionLocalService.getFileVersions(
-			fileEntry.getFileEntryId(), WorkflowConstants.STATUS_ANY);
-
-		for (DLFileVersion curDLFileVersion : dlFileVersions) {
-			curDLFileVersion.setStatus(WorkflowConstants.STATUS_IN_TRASH);
-
-			_dlFileVersionPersistence.update(curDLFileVersion);
-		}
-
-		if (!DLAppHelperThreadLocal.isEnabled()) {
-			return fileEntry;
-		}
-
-		DLFileEntry dlFileEntry = (DLFileEntry)fileEntry.getModel();
-
-		DLFileVersion oldDLFileVersion = (DLFileVersion)fileVersion.getModel();
-
-		int oldDLFileVersionStatus = oldDLFileVersion.getStatus();
-
-		TrashEntry trashEntry = _trashEntryLocalService.addTrashEntry(
-			userId, dlFileEntry.getGroupId(),
-			DLFileEntryConstants.getClassName(), dlFileEntry.getFileEntryId(),
-			dlFileEntry.getUuid(), dlFileEntry.getClassName(),
-			oldDLFileVersionStatus, dlFileVersionStatusOVPs,
-			UnicodePropertiesBuilder.put(
-				"fileName", dlFileEntry.getFileName()
-			).put(
-				"title", dlFileEntry.getTitle()
-			).build());
-
-		String trashTitle = _trashHelper.getTrashTitle(trashEntry.getEntryId());
-
-		dlFileEntry.setFileName(trashTitle);
-		dlFileEntry.setTitle(trashTitle);
-
-		dlFileEntry = _dlFileEntryPersistence.update(dlFileEntry);
-
-		// Indexer
-
-		Indexer<DLFileEntry> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
-			DLFileEntry.class);
-
-		indexer.reindex(dlFileEntry);
-
-		// Social
-
-		JSONObject extraDataJSONObject = JSONUtil.put(
-			"title", _trashHelper.getOriginalTitle(fileEntry.getTitle()));
-
-		SocialActivityManagerUtil.addActivity(
-			userId, fileEntry, SocialActivityConstants.TYPE_MOVE_TO_TRASH,
-			extraDataJSONObject.toString(), 0);
-
-		// Workflow
-
-		int oldStatus = fileVersion.getStatus();
-
-		if (oldStatus == WorkflowConstants.STATUS_PENDING) {
-			_workflowInstanceLinkLocalService.deleteWorkflowInstanceLink(
-				fileVersion.getCompanyId(), fileVersion.getGroupId(),
-				DLFileEntryConstants.getClassName(),
-				fileVersion.getFileVersionId());
-		}
-
-		return fileEntry;
-	}
-
-	protected Folder doMoveFolderFromTrash(
-			long userId, Folder folder, long parentFolderId,
-			ServiceContext serviceContext)
-		throws PortalException {
-
-		DLFolder dlFolder = (DLFolder)folder.getModel();
-
-		if (!dlFolder.isInTrash()) {
-			throw new RestoreEntryException(
-				RestoreEntryException.INVALID_STATUS);
-		}
-
-		if (_trashHelper.isInTrashExplicitly(dlFolder)) {
-			restoreFolderFromTrash(userId, folder);
-		}
-		else {
-
-			// Folder
-
-			TrashVersion trashVersion = _trashVersionLocalService.fetchVersion(
-				DLFolder.class.getName(), dlFolder.getFolderId());
-
-			int status = WorkflowConstants.STATUS_APPROVED;
-
-			if (trashVersion != null) {
-				status = trashVersion.getStatus();
-			}
-
-			_dlFolderLocalService.updateStatus(
-				userId, folder.getFolderId(), status,
-				new HashMap<String, Serializable>(), new ServiceContext());
-
-			// Trash
-
-			if (trashVersion != null) {
-				_trashVersionLocalService.deleteTrashVersion(trashVersion);
-			}
-
-			// Folders, file entries, and file shortcuts
-
-			restoreDependentsFromTrash(dlFolder);
-
-			// Sync
-
-			triggerRepositoryEvent(
-				folder.getRepositoryId(),
-				TrashRepositoryEventType.EntryRestored.class, Folder.class,
-				folder);
-
-			// Social
-
-			JSONObject extraDataJSONObject = JSONUtil.put(
-				"title", folder.getName());
-
-			SocialActivityManagerUtil.addActivity(
-				userId, folder, SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
-				extraDataJSONObject.toString(), 0);
-		}
-
-		return _dlAppLocalService.moveFolder(
-			userId, folder.getFolderId(), parentFolderId, serviceContext);
-	}
-
-	protected Folder doMoveFolderToTrash(long userId, Folder folder)
-		throws PortalException {
-
-		// Folder
-
-		DLFolder dlFolder = (DLFolder)folder.getModel();
-
-		if (dlFolder.isInTrash()) {
-			throw new TrashEntryException();
-		}
-
-		dlFolder = _dlFolderLocalService.updateStatus(
-			userId, folder.getFolderId(), WorkflowConstants.STATUS_IN_TRASH,
-			new HashMap<String, Serializable>(), new ServiceContext());
-
-		// Trash
-
-		TrashEntry trashEntry = _trashEntryLocalService.addTrashEntry(
-			userId, dlFolder.getGroupId(), DLFolderConstants.getClassName(),
-			dlFolder.getFolderId(), dlFolder.getUuid(), null,
-			WorkflowConstants.STATUS_APPROVED, null,
-			UnicodePropertiesBuilder.put(
-				"title", dlFolder.getName()
-			).build());
-
-		dlFolder.setName(_trashHelper.getTrashTitle(trashEntry.getEntryId()));
-
-		dlFolder = _dlFolderPersistence.update(dlFolder);
-
-		// Folders, file entries, and file shortcuts
-
-		moveDependentsToTrash(dlFolder);
-
-		// Sync
-
-		triggerRepositoryEvent(
-			folder.getRepositoryId(),
-			TrashRepositoryEventType.EntryTrashed.class, Folder.class, folder);
-
-		// Social
-
-		JSONObject extraDataJSONObject = JSONUtil.put(
-			"title", folder.getName());
-
-		SocialActivityManagerUtil.addActivity(
-			userId, folder, SocialActivityConstants.TYPE_MOVE_TO_TRASH,
-			extraDataJSONObject.toString(), 0);
-
-		// Indexer
-
-		Indexer<DLFolder> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
-			DLFolder.class);
-
-		indexer.reindex(dlFolder);
-
-		return new LiferayFolder(dlFolder);
-	}
-
 	protected ActionableDynamicQuery getCancelCheckOutsActionableDynamicQuery(
 		long groupId) {
 
@@ -1589,28 +871,6 @@ public class DLAppHelperLocalServiceImpl
 		return fileEntryActionableDynamicQuery;
 	}
 
-	protected List<ObjectValuePair<Long, Integer>> getDlFileVersionStatuses(
-		List<DLFileVersion> dlFileVersions) {
-
-		List<ObjectValuePair<Long, Integer>> dlFileVersionStatusOVPs =
-			new ArrayList<>(dlFileVersions.size());
-
-		for (DLFileVersion dlFileVersion : dlFileVersions) {
-			int status = dlFileVersion.getStatus();
-
-			if (status == WorkflowConstants.STATUS_PENDING) {
-				status = WorkflowConstants.STATUS_DRAFT;
-			}
-
-			ObjectValuePair<Long, Integer> dlFileVersionStatusOVP =
-				new ObjectValuePair<>(dlFileVersion.getFileVersionId(), status);
-
-			dlFileVersionStatusOVPs.add(dlFileVersionStatusOVP);
-		}
-
-		return dlFileVersionStatusOVPs;
-	}
-
 	protected long getFileEntryTypeId(FileEntry fileEntry) {
 		if (fileEntry instanceof LiferayFileEntry) {
 			DLFileEntry dlFileEntry = (DLFileEntry)fileEntry.getModel();
@@ -1634,286 +894,6 @@ public class DLAppHelperLocalServiceImpl
 			throw new IllegalArgumentException(
 				String.format("Unsupported sync event %s", syncEvent));
 		}
-	}
-
-	protected void trashOrRestoreFolder(DLFolder dlFolder, boolean moveToTrash)
-		throws PortalException {
-
-		TrashEntry trashEntry = null;
-
-		if (moveToTrash) {
-			trashEntry = _trashEntryLocalService.getEntry(
-				DLFolderConstants.getClassName(), dlFolder.getFolderId());
-		}
-
-		long dlFileEntryClassNameId = _classNameLocalService.getClassNameId(
-			DLFileEntry.class);
-
-		List<DLFileEntry> dlFileEntries = _dlFileEntryFinder.findByC_T(
-			dlFileEntryClassNameId, dlFolder.getTreePath());
-
-		for (DLFileEntry dlFileEntry : dlFileEntries) {
-			_assetEntryLocalService.updateVisible(
-				DLFileEntry.class.getName(), dlFileEntry.getFileEntryId(),
-				!moveToTrash);
-		}
-
-		long dlFolderClassNameId = _classNameLocalService.getClassNameId(
-			DLFolder.class);
-
-		List<DLFolder> dlFolders = _dlFolderFinder.findF_ByC_T(
-			dlFolderClassNameId, dlFolder.getTreePath());
-
-		for (DLFolder curDLFolder : dlFolders) {
-			_assetEntryLocalService.updateVisible(
-				DLFolder.class.getName(), curDLFolder.getFolderId(),
-				!moveToTrash);
-		}
-
-		if (moveToTrash) {
-			dlFolders = _dlFolderPersistence.findByG_M_LikeT_H_NotS(
-				dlFolder.getGroupId(), false,
-				CustomSQLUtil.keywords(
-					dlFolder.getTreePath(), WildcardMode.TRAILING)[0],
-				false, WorkflowConstants.STATUS_IN_TRASH);
-		}
-		else {
-			dlFolders = _dlFolderPersistence.findByG_M_LikeT_H(
-				dlFolder.getGroupId(), false,
-				CustomSQLUtil.keywords(
-					dlFolder.getTreePath(), WildcardMode.TRAILING)[0],
-				false);
-		}
-
-		if (!dlFolders.contains(dlFolder)) {
-			dlFolders = new ArrayList<>(dlFolders);
-
-			dlFolders.add(dlFolder);
-		}
-
-		for (DLFolder childDLFolder : dlFolders) {
-			trashOrRestoreFolder(
-				dlFolder, childDLFolder, moveToTrash, trashEntry);
-		}
-	}
-
-	protected void trashOrRestoreFolder(
-			DLFolder dlFolder, DLFolder childDLFolder, boolean moveToTrash,
-			TrashEntry trashEntry)
-		throws PortalException {
-
-		List<Long> dlFileEntryIds = new ArrayList<>();
-
-		List<DLFileEntry> dlFileEntries =
-			_dlFileEntryLocalService.getFileEntries(
-				childDLFolder.getGroupId(), childDLFolder.getFolderId());
-
-		for (DLFileEntry dlFileEntry : dlFileEntries) {
-			if (moveToTrash) {
-				if (_trashHelper.isInTrashExplicitly(dlFileEntry)) {
-					continue;
-				}
-			}
-			else if (!_trashHelper.isInTrashImplicitly(dlFileEntry)) {
-				continue;
-			}
-
-			// File shortcut
-
-			_dlFileShortcutLocalService.updateFileShortcutsActive(
-				dlFileEntry.getFileEntryId(), !moveToTrash);
-
-			// File versions
-
-			List<DLFileVersion> dlFileVersions = null;
-
-			if (moveToTrash) {
-				dlFileVersions = _dlFileVersionLocalService.getFileVersions(
-					dlFileEntry.getFileEntryId(), WorkflowConstants.STATUS_ANY);
-			}
-			else {
-				dlFileVersions = _dlFileVersionLocalService.getFileVersions(
-					dlFileEntry.getFileEntryId(),
-					WorkflowConstants.STATUS_IN_TRASH);
-			}
-
-			for (DLFileVersion dlFileVersion : dlFileVersions) {
-
-				// File version
-
-				if (moveToTrash) {
-					int oldStatus = dlFileVersion.getStatus();
-
-					dlFileVersion.setStatus(WorkflowConstants.STATUS_IN_TRASH);
-
-					dlFileVersion = _dlFileVersionPersistence.update(
-						dlFileVersion);
-
-					// Trash
-
-					if (oldStatus != WorkflowConstants.STATUS_APPROVED) {
-						int newStatus = oldStatus;
-
-						if (oldStatus == WorkflowConstants.STATUS_PENDING) {
-							newStatus = WorkflowConstants.STATUS_DRAFT;
-						}
-
-						_trashVersionLocalService.addTrashVersion(
-							trashEntry.getEntryId(),
-							DLFileVersion.class.getName(),
-							dlFileVersion.getFileVersionId(), newStatus, null);
-					}
-
-					// Workflow
-
-					if (oldStatus == WorkflowConstants.STATUS_PENDING) {
-						_workflowInstanceLinkLocalService.
-							deleteWorkflowInstanceLink(
-								dlFileVersion.getCompanyId(),
-								dlFileVersion.getGroupId(),
-								DLFileEntryConstants.getClassName(),
-								dlFileVersion.getFileVersionId());
-					}
-				}
-				else {
-					TrashVersion trashVersion =
-						_trashVersionLocalService.fetchVersion(
-							DLFileVersion.class.getName(),
-							dlFileVersion.getFileVersionId());
-
-					int oldStatus = WorkflowConstants.STATUS_APPROVED;
-
-					if (trashVersion != null) {
-						oldStatus = trashVersion.getStatus();
-					}
-
-					dlFileVersion.setStatus(oldStatus);
-
-					_dlFileVersionPersistence.update(dlFileVersion);
-
-					// Trash
-
-					if (trashVersion != null) {
-						_trashVersionLocalService.deleteTrashVersion(
-							trashVersion);
-					}
-				}
-			}
-
-			dlFileEntryIds.add(dlFileEntry.getFileEntryId());
-		}
-
-		if (!dlFileEntryIds.isEmpty()) {
-			dlAppHelperLocalService.reindex(
-				dlFolder.getCompanyId(), dlFileEntryIds);
-		}
-
-		List<DLFileShortcut> dlFileShortcuts =
-			_dlFileShortcutPersistence.findByG_F(
-				childDLFolder.getGroupId(), childDLFolder.getFolderId());
-
-		for (DLFileShortcut dlFileShortcut : dlFileShortcuts) {
-			if (moveToTrash) {
-				if (_trashHelper.isInTrashExplicitly(dlFileShortcut)) {
-					continue;
-				}
-
-				int oldStatus = dlFileShortcut.getStatus();
-
-				dlFileShortcut.setStatus(WorkflowConstants.STATUS_IN_TRASH);
-
-				dlFileShortcut = _dlFileShortcutPersistence.update(
-					dlFileShortcut);
-
-				// Trash
-
-				if (oldStatus != WorkflowConstants.STATUS_APPROVED) {
-					_trashVersionLocalService.addTrashVersion(
-						trashEntry.getEntryId(),
-						DLFileShortcutConstants.getClassName(),
-						dlFileShortcut.getFileShortcutId(), oldStatus, null);
-				}
-			}
-			else {
-				if (!_trashHelper.isInTrashImplicitly(dlFileShortcut)) {
-					continue;
-				}
-
-				TrashVersion trashVersion =
-					_trashVersionLocalService.fetchVersion(
-						DLFileShortcutConstants.getClassName(),
-						dlFileShortcut.getFileShortcutId());
-
-				int oldStatus = WorkflowConstants.STATUS_APPROVED;
-
-				if (trashVersion != null) {
-					oldStatus = trashVersion.getStatus();
-				}
-
-				dlFileShortcut.setStatus(oldStatus);
-
-				_dlFileShortcutPersistence.update(dlFileShortcut);
-
-				if (trashVersion != null) {
-					_trashVersionLocalService.deleteTrashVersion(trashVersion);
-				}
-			}
-		}
-
-		if (childDLFolder.equals(dlFolder)) {
-			return;
-		}
-
-		if (moveToTrash) {
-			if (_trashHelper.isInTrashExplicitly(childDLFolder)) {
-				return;
-			}
-
-			int oldStatus = childDLFolder.getStatus();
-
-			childDLFolder.setStatus(WorkflowConstants.STATUS_IN_TRASH);
-
-			childDLFolder = _dlFolderPersistence.update(childDLFolder);
-
-			// Trash
-
-			if (oldStatus != WorkflowConstants.STATUS_APPROVED) {
-				_trashVersionLocalService.addTrashVersion(
-					trashEntry.getEntryId(), DLFolder.class.getName(),
-					childDLFolder.getFolderId(), oldStatus, null);
-			}
-		}
-		else {
-			if (!_trashHelper.isInTrashImplicitly(childDLFolder)) {
-				return;
-			}
-
-			TrashVersion trashVersion = _trashVersionLocalService.fetchVersion(
-				DLFolder.class.getName(), childDLFolder.getFolderId());
-
-			int oldStatus = WorkflowConstants.STATUS_APPROVED;
-
-			if (trashVersion != null) {
-				oldStatus = trashVersion.getStatus();
-			}
-
-			childDLFolder.setStatus(oldStatus);
-
-			childDLFolder = _dlFolderPersistence.update(childDLFolder);
-
-			// Trash
-
-			if (trashVersion != null) {
-				_trashVersionLocalService.deleteTrashVersion(trashVersion);
-			}
-		}
-
-		// Indexer
-
-		Indexer<DLFolder> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
-			DLFolder.class);
-
-		indexer.reindex(childDLFolder);
 	}
 
 	protected <T extends RepositoryModel<T>> void triggerRepositoryEvent(
@@ -1954,11 +934,6 @@ public class DLAppHelperLocalServiceImpl
 			DLFileEntryConstants.getClassName(), fileEntryId);
 	}
 
-	private static volatile TrashHelper _trashHelper =
-		ServiceProxyFactory.newServiceTrackedInstance(
-			TrashHelper.class, DLAppHelperLocalServiceImpl.class,
-			"_trashHelper", false);
-
 	@BeanReference(type = AssetCategoryLocalService.class)
 	private AssetCategoryLocalService _assetCategoryLocalService;
 
@@ -1971,23 +946,11 @@ public class DLAppHelperLocalServiceImpl
 	@BeanReference(type = AssetTagLocalService.class)
 	private AssetTagLocalService _assetTagLocalService;
 
-	@BeanReference(type = ClassNameLocalService.class)
-	private ClassNameLocalService _classNameLocalService;
-
-	@BeanReference(type = DLAppLocalService.class)
-	private DLAppLocalService _dlAppLocalService;
-
 	@BeanReference(type = DLAppService.class)
 	private DLAppService _dlAppService;
 
-	@BeanReference(type = DLFileEntryFinder.class)
-	private DLFileEntryFinder _dlFileEntryFinder;
-
 	@BeanReference(type = DLFileEntryLocalService.class)
 	private DLFileEntryLocalService _dlFileEntryLocalService;
-
-	@BeanReference(type = DLFileEntryPersistence.class)
-	private DLFileEntryPersistence _dlFileEntryPersistence;
 
 	@BeanReference(type = DLFileShortcutLocalService.class)
 	private DLFileShortcutLocalService _dlFileShortcutLocalService;
@@ -1995,34 +958,11 @@ public class DLAppHelperLocalServiceImpl
 	@BeanReference(type = DLFileShortcutPersistence.class)
 	private DLFileShortcutPersistence _dlFileShortcutPersistence;
 
-	@BeanReference(type = DLFileVersionLocalService.class)
-	private DLFileVersionLocalService _dlFileVersionLocalService;
-
 	@BeanReference(type = DLFileVersionPersistence.class)
 	private DLFileVersionPersistence _dlFileVersionPersistence;
 
-	@BeanReference(type = DLFolderFinder.class)
-	private DLFolderFinder _dlFolderFinder;
-
-	@BeanReference(type = DLFolderLocalService.class)
-	private DLFolderLocalService _dlFolderLocalService;
-
-	@BeanReference(type = DLFolderPersistence.class)
-	private DLFolderPersistence _dlFolderPersistence;
-
 	@BeanReference(type = RatingsStatsLocalService.class)
 	private RatingsStatsLocalService _ratingsStatsLocalService;
-
-	@BeanReference(type = TrashEntryLocalService.class)
-	@SuppressWarnings("deprecation")
-	private TrashEntryLocalService _trashEntryLocalService;
-
-	@BeanReference(type = TrashVersionLocalService.class)
-	@SuppressWarnings("deprecation")
-	private TrashVersionLocalService _trashVersionLocalService;
-
-	@BeanReference(type = WorkflowInstanceLinkLocalService.class)
-	private WorkflowInstanceLinkLocalService _workflowInstanceLinkLocalService;
 
 	/**
 	 * @see com.liferay.document.library.sync.constants.DLSyncConstants
